break up integration specs into smaller files
