+++ spec/integration/an_organizer_with_after_each_callbacks_spec.rb
@@ -1,0 +1,34 @@
+# frozen_string_literal: true
+
+require 'spec_helper'
+
+RSpec.describe 'An organizer with .after_each callbacks', type: :integration do
+  let!(:interactor1) { build_interactor('TestInteractor1') }
+  let!(:interactor2) { build_interactor('TestInteractor2') }
+  let(:interactor_class) do
+    build_organizer do
+      after_each_perform :test_after_each_perform
+      organize TestInteractor1, TestInteractor2
+
+      private
+
+      def test_after_each_perform; end
+    end
+  end
+
+  include_examples 'a class with interactor methods'
+  include_examples 'a class with interactor callback methods'
+  include_examples 'a class with interactor context methods'
+  include_examples 'a class with organizer callback methods'
+
+  describe '.perform' do
+    subject { interactor_class.perform }
+
+    it { is_expected.to be_a interactor_class.context_class }
+    it 'is expected to receive #test_after_each_perform twice' do
+      expect_any_instance_of(interactor_class).to receive(:test_after_each_perform)
+        .exactly(:twice)
+      subject
+    end
+  end
+end
